[Ref][Mod]: Split reducer to middleware; added TODOs; start of reset logic
Originally was adding the resetRows logic when I found
some massive inefficiencies in doing the pairs the way
I was doing it. It's far too repetitive & chokes on big
sets.

Found some solutions but need to WIP that code as comment, as
I need to refactor how the Pairs work.